Move copilot auth into integrations module
Extract GitHub device flow HTTP logic into services/copilot_oauth.py,
consolidate schemas into integrations.py, and serve endpoints under
/integrations/copilot/* for consistency with the Gmail integration pattern.
